Add requirements list to specs
The enhancement is going to require changes to update_quality(), which is a bit of a mess. 

The best route forward when we implement the new conjured items feature is going to require some refactoring, but we can't refactor without tests. 

Added in pending tests to cover all the current system requirements.